Revert "Add files via upload (#4)"
This reverts commit e414ea8b21bcf9d042d6020eea4d446b23cde039.
