+++ src/lstm/lstm.py
@@ -1,0 +1,16 @@
+import torch.nn as nn
+
+class lstm_model(nn.Module):
+    def __init__(self, input_dim, hidden_dim, layer_dim, output_dim):
+        super(lstm_model, self).__init__()
+        
+        self.hidden_dim = hidden_dim
+        self.layer_dim = layer_dim
+
+        self.lstm = nn.LSTM(input_size=input_dim, hidden_size=hidden_dim, num_layers=layer_dim, batch_first=True)
+        self.linear = nn.Linear(hidden_dim, output_dim)
+
+    def forward(self, x):
+        x, _ = self.lstm(x)
+        x = self.linear(x)
+        return x
